Update utils.py to use trace() from pdb
Fixed the issue wit the pdb set_trace() when I wanted to turn tracing
off and on. It now works the way I want it to.

Updated the base_config.yaml file to just use true/false without the
dictionary to designate if an object is in the set or not.