Update boundbuilder.py to build without config
Update boudnbuilder.py to build from command line without a config.
There is a lot of duplicate code here so it is not really usable but it
functions.

I will be deduplicating the code and doing some major refactoring soon
and removing the support for .smpl files, since I am moving to YAML from
here on out.